chore: fastapi refactoring
